+++ src/main/java/com/pierrette/api/services/LocalisationService.java
@@ -1,7 +1,9 @@
 package com.pierrette.api.services;
 
 import com.pierrette.api.entities.Localisation;
+import com.pierrette.api.entities.Operateur;
 import com.pierrette.api.repositories.LocalisationRepo;
+import com.pierrette.api.repositories.OperateurRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -13,19 +15,37 @@ public class LocalisationService {
     @Autowired
     private LocalisationRepo localisationRepository;
 
-    public List<Localisation> findAll() {
+
+    public List<Localisation> getAllLocalisations(){
         return localisationRepository.findAll();
     }
 
-    public Optional<Localisation> findById(Integer id) {
-        return localisationRepository.findById(id);
-    }
-
-    public Localisation save(Localisation localisation) {
+    public Localisation createLocalisation(Localisation localisation){
         return localisationRepository.save(localisation);
     }
 
-    public void deleteById(Integer id) {
-        localisationRepository.deleteById(id);
+    public Localisation updateLocalisation(Integer id,Localisation localisationUpdate){
+        Localisation localisationExiste = localisationRepository.findById(id).orElse(null);
+        if (localisationExiste!=null){
+            localisationExiste.setLatitude(localisationUpdate.getLatitude());
+            localisationExiste.setLongitude(localisationUpdate.getLongitude());
+
+            Localisation localisationAjour = localisationRepository.save(localisationExiste);
+
+        } return localisationExiste;
     }
+
+    public String deleteLocalisation(Integer id) {
+        if(id!=0){
+            localisationRepository.deleteById(id);
+            return "Suppression";
+        }
+        return null;
+    }
+
+    public Localisation getLocalisation(Integer idLocalisation){
+        return localisationRepository.findById(idLocalisation).get();
+    }
+
 }
+
